chore: apply @mini-max changes for PR #9128
