+++ src/core/resource-directory.json
@@ -1,3 +1,3 @@
 {
-  "resourceDirectory": "/Users/henrykirk/GMLoop/resources"
+  "resourceDirectory": "/home/runner/work/GMLoop/GMLoop/resources"
 }
